+++ hc/api/tests/test_list_checks.py
@@ -17,8 +17,7 @@ class ListChecksTestCase(BaseTestCase):
         self.a1 = Check(user=self.alice, name="Alice 1")
         self.a1.timeout = td(seconds=3600)
         self.a1.grace = td(seconds=900)
-        self.a1.last_ping = self.now
-        self.a1.n_pings = 1
+        self.a1.n_pings = 0
         self.a1.status = "new"
         self.a1.tags = "a1-tag a1-additional-tag"
         self.a1.save()
@@ -45,19 +44,16 @@ class ListChecksTestCase(BaseTestCase):
         doc = r.json()
         self.assertEqual(len(doc["checks"]), 2)
 
-        a1 = None
-        a2 = None
+        by_name = {}
         for check in doc["checks"]:
-            if check["name"] == "Alice 1":
-                a1 = check
-            if check["name"] == "Alice 2":
-                a2 = check
+            by_name[check["name"]] = check
 
+        a1 = by_name["Alice 1"]
         self.assertEqual(a1["timeout"], 3600)
         self.assertEqual(a1["grace"], 900)
         self.assertEqual(a1["ping_url"], self.a1.url())
-        self.assertEqual(a1["last_ping"], self.now.isoformat())
-        self.assertEqual(a1["n_pings"], 1)
+        self.assertEqual(a1["last_ping"], None)
+        self.assertEqual(a1["n_pings"], 0)
         self.assertEqual(a1["status"], "new")
         self.assertEqual(a1["channels"], str(self.c1.code))
 
@@ -66,13 +62,16 @@ class ListChecksTestCase(BaseTestCase):
         self.assertEqual(a1["update_url"], update_url)
         self.assertEqual(a1["pause_url"], pause_url)
 
-        next_ping = self.now + td(seconds=3600)
-        self.assertEqual(a1["next_ping"], next_ping.isoformat())
+        self.assertEqual(a1["next_ping"], None)
 
+        a2 = by_name["Alice 2"]
         self.assertEqual(a2["timeout"], 86400)
         self.assertEqual(a2["grace"], 3600)
         self.assertEqual(a2["ping_url"], self.a2.url())
         self.assertEqual(a2["status"], "up")
+        next_ping = self.now + td(seconds=86400)
+        self.assertEqual(a2["last_ping"], self.now.isoformat())
+        self.assertEqual(a2["next_ping"], next_ping.isoformat())
 
     def test_it_handles_options(self):
         r = self.client.options("/api/v1/checks/")